Added UI basic events, main menu state, map editor states[WIP]

+++ src/UI/Components/C_UIBackground.cpp
@@ -40,7 +40,7 @@ namespace rts
                 m_background.setTexture( *ResourceManager::getTexture( texID ) );
                 
                 // NOTE: This line has to be present in the code which handles updating the Background
-                m_background.setTextureRect( sf::IntRect{ m_state * m_sWidth, 0, m_sWidth, m_sHeight } );
+                m_background.setTextureRect( sf::IntRect{ static_cast<int>( m_state ) * m_sWidth, 0, m_sWidth, m_sHeight } );
             }
             
             void C_UIBackground::draw(sf::RenderTarget& target, sf::RenderStates states) const
